submissions/miscellaneous/heet_linguaflex/requirements.txt
alembic==1.12.0
anyio==4.0.0
async-timeout==4.0.3
bidict==0.22.1
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.3.0
click==8.1.7
cloudpickle==2.2.1
distro==1.8.0
exceptiongroup==1.1.3
fastapi==0.96.1
googletrans==3.1.0a0
greenlet==3.0.0
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
install==1.3.5
Jinja2==3.1.2
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
packaging==23.2
Pillow==10.0.1
platformdirs==3.11.0
psutil==5.9.5
pydantic==1.10.13
Pygments==2.16.1
python-engineio==4.7.1
python-multipart==0.0.5
python-socketio==5.9.0
redis==4.6.0
reflex==0.2.9
requests==2.31.0
rfc3986==1.5.0
rich==13.6.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.8
starlette==0.27.0
starlette-admin==0.9.0
typer==0.4.2
typing_extensions==4.8.0
urllib3==2.0.6
uvicorn==0.20.0
watchdog==2.3.1
watchfiles==0.19.0
websockets==10.4
wrapt==1.15.0
wsproto==1.2.0
